coffeeshop/requirements.txt
openai==1.12.0
python-dotenv==1.0.1
